cpu: Differentiate between threads on code/mem logging
Without this, when all threads are set to log, you don't
know what logging line belongs to what thread. Each running
thread has a different unicorn instance associated with it
so it can be used to differentiate between them.

In the future, we'd save the thread id on TLS and recover
it from there, but this works for now.