modules/SceTouch: add support for multi-touch input on touchscreen devices
Up to 8 fingers are registered but not at the same time.
In theory the vita should only acept 6 on the front and 4 on the back.
Force imput is also transfered if the device supports it.

The user can toggle between front touchscreen or back touchpad with the
"T" key.
The key has also been added to the controls dialog in the gui.