shader: fix bugs in loading and storing of vertex output with non-four components
First bug: When the component size is not 4, the vertext shader will create out variable with the actual size (that is not 4), but fragment shader always uses 4 for input variable size.
Second bug: Wen the copmponent size is 1, fragment shader will load the variable as vector with size 4. (since it's not finalized) However, util::store uses the spirv function to get components to write (it doesn't utilize dest_mask for determining size)

With these bugs fixed, we can safely remove the assert(number_of_comp_vec != 1) that we had for a long time.