Updating Spanish translation, splitting into two
Greetings.

Here is an update for the Spanish translation of Vita3K (well, at least the part that is translateable, damn ImGUI, it's ubiquity, simplicity lack of localization features!) that adds the new strings but also does something a bit different: splits the translation into two distinct regions of Spanish: European Spanish (es.xml) and Latin American Spanish (es-419.xml).

This will probably need some changes from the code to work correctly. The changes are minimal, but since there was already a mix of variants in the old file and the new one contains one particular element (a controller of the gamepad type) that has different names depending on the side of the pond, I wanted to make use of this time to split the translation depending on the speakers. That way, any future translators will be funneled to their respective variants and will not mix them.