kernel: don't check status on msgpipe finish
The "finish()" lambda expected the thread to be in waiting state and checked this using the 'expected' argument of thread->update_status().
However, the wakeup of threads sleeping on a message pipe is done by changing their thread from wait to run, breaking this assumption.
This leads to an assertion failure and crash, preventing e.g. Ratchet & Clank from booting.

Fix (workaround) by no longer verifying the thread's status on finish.