Implement button rebinds
This PR is an attempt at implementing rebinds for controller buttons. Binding buttons to sticks (which, I imagine would also be requested by some) is not implemented here.

As this is my first (serious) commit to this project, feedback is greatly appreciated since I'm almost certain some of this is sub-par.