ctrl: Drop inputs while rebinding menu is open
Make it so that when the rebinds menu is open that it does not affect the GUI and in-game events.

Currently I have it set up so that events get dropped when the controller BIND menu is up, and not the controller LIST menu, but this can be easily changed.